requirements.txt
# Core Framework
Flask==3.1.2
Werkzeug==3.1.1

# Database
pymongo==4.15.0

# AI/LLM
langchain-ollama==0.3.8
langchain-core==0.3.76

# Environment & Configuration
python-dotenv==1.1.1

# Testing
pytest==8.4.2
pytest-mock==3.15.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Optional: For future UI development
# flask-cors==5.0.0
# flask-restx==1.3.0